+++ contracts/interfaces/pool/ICrocSwapPoolState.sol
@@ -27,7 +27,7 @@ interface ICrocSwapPoolState {
 
     /// @notice The amounts of token0 and token1 that are owed to the protocol
     /// @dev Protocol fees will never exceed uint128 max in either token
-    function protocolFees() external view returns (uint128 token0, uint128 token1);
+    function protocolFees() external view returns (uint256 token0, uint256 token1);
 
     /// @notice The currently in range liquidity available to the pool
     /// @dev This value has no relationship to the total liquidity across all ticks
